+++ src/client/js/set_options.js
@@ -9,7 +9,17 @@ import { AccountsUI } from 'meteor/pwix:accounts-ui';
 import { FlowRouter } from 'meteor/ostrio:flow-router-extra';
 import { Tracker } from 'meteor/tracker';
 
-// ask to AccountsUI to push the onVerifiedEmail configuration, replacing with the one provided
+const _parms = [
+    'onEmailVerifiedBeforeFn',
+    'onEmailVerifiedBox',
+    'onEmailVerifiedBoxTitle',
+    'onEmailVerifiedBoxMessage',
+    'onEmailVerifiedBoxCb'
+];
+
+let _values = null;
+
+// ask to AccountsUI to push the onEmailVerified configuration, replacing with the one provided
 //  the configuration will be restored as soon as we have got an administrator
 SAA._setOptions = function(){
     // callback function
@@ -17,11 +27,18 @@ SAA._setOptions = function(){
         FlowRouter.go( '/' );
     };
     // code
-    AccountsUI.saveOnce( 'onVerifiedEmail' );
+    if( !_values ){
+        _values = {};
+        _parms.every(( p ) => {
+            _values[p] = AccountsUI.opts()[p]();
+        });
+    }
+    //AccountsUI.saveOnce( _parms );
     // override the pwix:accounts-ui configuration to display our own message when there will be a new admin
-    AccountsUI.opts().onVerifiedEmailTitle({ namespace: I18N, i18n: 'confirm.title' });
-    AccountsUI.opts().onVerifiedEmailMessage({ namespace: I18N, i18n: 'confirm.permsgot' });
-    AccountsUI.opts().onVerifiedEmailCb( _cb );
+    AccountsUI.opts().onEmailVerifiedBox( true );
+    AccountsUI.opts().onEmailVerifiedTitle({ namespace: I18N, i18n: 'confirm.title' });
+    AccountsUI.opts().onEmailVerifiedMessage({ namespace: I18N, i18n: 'confirm.permsgot' });
+    AccountsUI.opts().onEmailVerifiedAfterFn( _cb );
 };
 
 // wait for the package configuration be done
@@ -29,7 +46,7 @@ Meteor.startup(() => {
     if( SAA._conf.verbosity & SAA.C.Verbose.STATUS ){
         console.log( 'pwix:startup-app-admin SAA._conf.requireVerifiedEmail', SAA._conf.requireVerifiedEmail, 'countAdmins', SAA.countAdmins.get());
     }
-    if( SAA._conf.requireVerifiedEmail && SAA.countAdmins.get() <= 0 ){
+    if( SAA._conf.requireVerifiedEmail && SAA.countAdmins.get() === 0 ){
         SAA._setOptions();
     }
 });
@@ -37,7 +54,12 @@ Meteor.startup(() => {
 // restore the original pwix:accounts-ui configuration as soon as we get an administrator
 Tracker.autorun(() => {
     if( SAA.countAdmins.get() > 0 && SAA.waitForEmailVerification()){
-        AccountsUI.restore( 'onVerifiedEmail' );
+        //AccountsUI.restore( _parms );
+        if( _values ){
+            _parms.every(( p ) => {
+                AccountsUI.opts()[p]( _values[p] );
+            });
+        }
         SAA.waitForEmailVerification( false );
     }
 });