+++ ChangeLog.md
@@ -19,6 +19,7 @@
     - Make the package auto-align itself inside of the provided parent div
     - Change the default admin role to APP_ADMINISTRATOR
     - Meteor requirement is back to 2.9.0
+    - Bump pwix:accounts-ui requirement to v 1.4.0
 
 ### 1.1.1